--- data/errors/__pycache__/internals/__init__.py
@@ -1,1 +1,0 @@
-from .text import sql_like_to_regex, build_regex
