--- data/expectations/__init__.py
@@ -1,2 +1,0 @@
-from .expect import Expectations
-from .evaluate import test_record
